requirements.txt
streamlit>=1.28.0
ollama>=0.1.0
Pillow>=9.0.0